# labs/lab2/openapi.yaml
openapi: 3.1.1
info:
  title: FindMe API
  version: v1
servers:
  - url: http://localhost:8080
    description: Local server
paths:
...
  /api/faq:
    get:
      tags:
        - faq-controller
      operationId: getAllFaqs
      responses:
        "200":
          description: List of FAQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Faq'
    post:
      tags:
        - faq-controller
      operationId: createFaq
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Faq'
      responses:
        "201":
          description: FAQ created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Faq'
...
  /api/faq/{id}:
    get:
      tags:
        - faq-controller
      operationId: getFaqById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: FAQ found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Faq'
        "404":
          description: FAQ not found
    delete:
      tags:
        - faq-controller
      operationId: deleteFaq
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: FAQ deleted
        "404":
          description: FAQ not found
    patch:
      tags:
        - faq-controller
      operationId: updateFaqQuestion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                question:
                  type: string
              required:
                - question
      responses:
        "200":
          description: FAQ updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Faq'
        "404":
          description: FAQ not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Faq:
      type: object
      properties:
        id:
          type: integer
          format: int64
        question:
          type: string
        answer:
          type: string
        userId:
          type: integer
          format: int64
...